refactored analysis to package
